birth year to int from float

+++ bikeshare.py
@@ -8,6 +8,7 @@ new_york = pd.read_csv("./new_york_city.csv")
 chicago = pd.read_csv("./chicago.csv")
 washington = pd.read_csv("./washington.csv")
 cities = [new_york,chicago,washington]
+
 # Lets add the missing columns for washington
 washington['Gender'] = np.nan
 washington['Birth Year'] = pd.Series(dtype='Int64')
@@ -18,7 +19,7 @@ washington['Trip Duration'] = washington['Trip Duration'].round(0).astype('Int64
 for city in cities:
     city.drop(columns='Unnamed: 0', inplace=True)
     city['Start Time'] = pd.to_datetime(city['Start Time'])
-    #city['Birth Year'] = city['Birth Year'].astype('Int64')
+    city['Birth Year'] = city['Birth Year'].astype('Int64')
     city['Hour'] = city['Start Time'].dt.strftime('%I %p').str.lstrip('0')
     city['Month'] = city['Start Time'].dt.month_name().str.upper()
     city['Day'] = city['Start Time'].dt.day_name().str.upper()
